codegen/pyproject.toml
[tool]
[tool.poetry]
name = "codegen"
version = "0.0.1"
description = "Top-level package for codegen."
authors = ["Florian d'Erfurth <florian@derfurth.com>", "Fanny Cheung <fanny.cheung@beta.gouv.fr>"]
readme = "README.md"
license = "todo"
classifiers = [
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3.9',
]
packages = [
    { include = "codegen" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = "^3.9"

# Templating
Jinja2 = "^2.11.2"

# Formatting generated files
black = "^20.8b1"
jsbeautifier = "^1.13.5"
beautifulsoup4 = "^4.9.3"

# Parsing
mistletoe = "^0.7.2"
PyYAML = "^5.3.1"
python-docx = "^0.8.10"
pandas = "^1.2.1"
openpyxl = "^3.0.6"

# Command line
typer = "^0.3.2"

[tool.poetry.dev-dependencies]
bumpversion = "*"
coverage = {extras = ["toml"], version = "^5.4"}
flake8 = "*"
invoke = "*"
isort = "*"
pylint = "*"
pytest = "*"
yapf = "*"
pytest-cov = "^2.11.1"

[tool.poetry.scripts]
extract = 'codegen.cli_extract:app'
generate = 'codegen.cli_generate:app'
regenerate = 'codegen.cli_regenerate:app'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.coverage.run]
branch = true
source = ["codegen"]

[tool.coverage.report]
show_missing = true
